Move secure partition interface from cli/src to lib
