+++ lib/secure_partition_interface.c
