+++ lib/secure_partition_interface.h
